Move mainImage state. Move Image type definition.
