+++ src/types/index.ts
@@ -1,0 +1,6 @@
+export interface Image {
+  blob: string
+  name: string
+  timeStamp: string
+  fileData: FileSystemFileEntry
+}
